Handle Content-Length values above 2 GiB in the proxy
The request framing check parsed Content-Length as a long, but the code that
decides how to read the body and the relay itself used an int. A value of
2^31 or more was therefore read as "no Content-Length": a request was
classified as having no body, so its body bytes were parsed as the next
request on that connection, and a response fell back to "body ends when the
connection closes", which a keep-alive upstream never does.

Content-Length is now carried as a long from parsing to relaying: HeaderLines
parses it with getHeaderValueAsLong, HttpRequest/HttpResponse hold a Long, and
RelayContentLength counts remaining bytes in a long. Values beyond the range of
a long are still rejected by the framing check.

Tests cover request and response classification, rejection beyond a long, and
relaying a body of 2 GiB + 5000 bytes.

+++ src/main/java/tricatch/oe/proxy/http/io/HttpRequest.java
@@ -11,7 +11,7 @@ public class HttpRequest {
     private final String version;
     private final String host;
     private final String connection;
-    private final Integer contentLength;
+    private final Long contentLength;
     private final HttpStream httpStream;
     private final HeaderLines headers;
     
@@ -26,7 +26,7 @@ public class HttpRequest {
      * @param httpStream body stream type
      * @param headers request headers
      */
-    public HttpRequest(String method, String path, String version, String host, String connection, Integer contentLength, HttpStream httpStream, HeaderLines headers) {
+    public HttpRequest(String method, String path, String version, String host, String connection, Long contentLength, HttpStream httpStream, HeaderLines headers) {
         this.method = method;
         this.path = path;
         this.version = version;
@@ -81,7 +81,7 @@ public class HttpRequest {
      * Get content length value
      * @return content length value
      */
-    public Integer getContentLength() {
+    public Long getContentLength() {
         return contentLength;
     }
